+++ phirSOFT.ConstraintedComparisonTests/UnitTest1.cs
@@ -11,7 +11,7 @@ using static System.String;
 namespace phirSOFT.ConstraintedComparisonTests
 {
     [TestClass]
-    public class UnitTest1
+    public partial class UnitTest1
     {
         [TestMethod]
         public void BehaviourTest01()
@@ -236,102 +236,16 @@ namespace phirSOFT.ConstraintedComparisonTests
             
         }
 
-        private class Circle : ITopologicalComparable<Circle>
+        [TestMethod]
+        public void TestInsertLast()
         {
-            public readonly int m;
-            private readonly int mod;
+            var items = new[] {3, 2, 1};
+            var list = new List<int>();
 
-            public Circle(int m, int mod)
+            foreach (var item in items)
             {
-                this.m = m;
-                this.mod = mod;
+                list.Insert(item);
             }
-
-            public int CompareTo(Circle other)
-            {
-                var tmp = m - other.m;
-                if (tmp < 0)
-                    tmp += mod;
-
-                if (tmp > mod / 2)
-                    tmp = -(mod - tmp);
-
-                return tmp;
-            }
-
-            public bool CanCompareTo(Circle other)
-            {
-                return mod == other.mod;
-            }
-
-            public override string ToString()
-            {
-                return $"{m} mod {mod}";
-            }
-        }
-
-    }
-
-    internal class AbsoluteTopological<T> : ITopologicalComparable<AbsoluteTopological<T>>
-    {
-        public T Value { get; }
-
-        private static readonly Dictionary<T, AbsoluteTopological<T>> Constraints = new Dictionary<T, AbsoluteTopological<T>>();
-
-        private readonly List<AbsoluteTopological<T>> _before = new List<AbsoluteTopological<T>>();
-
-        private readonly List<AbsoluteTopological<T>> _after = new List<AbsoluteTopological<T>>();
-
-
-        public AbsoluteTopological(T value)
-        {
-            Value = value;
-            Constraints.Add(value, this);
-        }
-
-
-        public void Before(AbsoluteTopological<T> other)
-        {
-            other._after.Add(this);
-            _before.Add(other);
-        }
-
-        public void After(AbsoluteTopological<T> other)
-        {
-            other._before.Add(this);
-            _after.Add(other);
-        }
-
-        public int CompareTo(AbsoluteTopological<T> other)
-        {
-            if (_after.Contains(other))
-                return 1;
-            if (_before.Contains(other))
-                return -1;
-            return 0;
-        }
-
-        public bool CanCompareTo(AbsoluteTopological<T> other)
-        {
-            return _after.Contains(other) || _before.Contains(other);
-        }
-
-        public static implicit operator T(AbsoluteTopological<T> val)
-        {
-            return val.Value;
-        }
-
-        public static implicit operator AbsoluteTopological<T>(T val)
-        {
-            if (Constraints.ContainsKey(val))
-                return Constraints[val];
-            var c = new AbsoluteTopological<T>(val);
-            return c;
-        }
-
-        public override string ToString()
-        {
-            return Value.ToString();
         }
     }
 }
